Move config code to own package
Moving config code to its own package allows third party code to import the
package and use upstream options parsing instead of implementing it themselves.
